Move provider into components
