+++ src/App.js
@@ -6,7 +6,7 @@ import GlobalStyles from 'src/components/GlobalStyles';
 import 'src/mixins/chartjs';
 import theme from 'src/theme';
 import routes from 'src/routes';
-import { BackendAPIProvider } from './api/BackendAPIProvider';
+import { BackendAPIProvider } from './components/BackendAPIProvider';
 
 const App = () => {
   const routing = useRoutes(routes);
